+++ app/Http/Controllers/ServiceTemplateController.php
@@ -5,7 +5,6 @@ namespace App\Http\Controllers;
 use App\Http\Interfaces\ToastInterface;
 use App\Models\Device;
 use App\Models\DeviceGroup;
-use App\Models\Service;
 use App\Models\ServiceTemplate;
 use Illuminate\Http\Request;
 use Illuminate\Validation\Rule;
@@ -14,11 +13,6 @@ use LibreNMS\Services;
 
 class ServiceTemplateController extends Controller
 {
-    public function __construct()
-    {
-        $this->authorizeResource(ServiceTemplate::class, 'template');
-    }
-
     /**
      * Display a listing of the resource.
      *
@@ -26,7 +20,7 @@ class ServiceTemplateController extends Controller
      */
     public function index()
     {
-        //$this->authorize('manage', ServiceTemplate::class);
+        $this->authorize('viewAny', ServiceTemplate::class);
 
         return view(
             'service-template.index', [
@@ -44,6 +38,8 @@ class ServiceTemplateController extends Controller
      */
     public function create()
     {
+        $this->authorize('create', ServiceTemplate::class);
+
         return view(
             'service-template.create', [
                 'template' => new ServiceTemplate(),
@@ -62,6 +58,8 @@ class ServiceTemplateController extends Controller
      */
     public function store(Request $request, ToastInterface $toast)
     {
+        $this->authorize('create', ServiceTemplate::class);
+
         $this->validate(
             $request, [
                 'name' => 'required|string|unique:service_templates',
@@ -81,7 +79,7 @@ class ServiceTemplateController extends Controller
             ]
         );
 
-        $template = ServiceTemplate::make(
+        $template = new ServiceTemplate(
             $request->only(
                 [
                     'name',
@@ -118,6 +116,8 @@ class ServiceTemplateController extends Controller
      */
     public function show(ServiceTemplate $template)
     {
+        $this->authorize('view', $template);
+
         return redirect(url('/services/templates/' . $template->id));
     }
 
@@ -129,6 +129,8 @@ class ServiceTemplateController extends Controller
      */
     public function edit(ServiceTemplate $template)
     {
+        $this->authorize('update', $template);
+
         return view(
             'service-template.edit', [
                 'template' => $template,
@@ -148,6 +150,8 @@ class ServiceTemplateController extends Controller
      */
     public function update(Request $request, ServiceTemplate $template, ToastInterface $toast)
     {
+        $this->authorize('update', $template);
+
         $this->validate(
             $request, [
                 'name' => [
@@ -232,9 +236,8 @@ class ServiceTemplateController extends Controller
      * Apply specified Service Template to Device Groups.
      *
      * @param  ServiceTemplate  $template
-     * @return \Illuminate\Http\RedirectResponse|\Illuminate\Http\Response|\Illuminate\View\View
      */
-    public function applyDeviceGroups(ServiceTemplate $template)
+    private function applyDeviceGroups(ServiceTemplate $template): void
     {
         foreach (DeviceGroup::inServiceTemplate($template->id)->get() as $device_group) {
             foreach (Device::inDeviceGroup($device_group->id)->get() as $device) {
@@ -255,18 +258,14 @@ class ServiceTemplateController extends Controller
                 );
             }
         }
-        $msg = __('Services for Template :name have been updated', ['name' => $template->name]);
-
-        return response($msg, 200);
     }
 
     /**
      * Apply specified Service Template to Devices.
      *
      * @param  ServiceTemplate  $template
-     * @return \Illuminate\Http\RedirectResponse|\Illuminate\Http\Response|\Illuminate\View\View
      */
-    public function applyDevices(ServiceTemplate $template)
+    private function applyDevices(ServiceTemplate $template): void
     {
         foreach (Device::inServiceTemplate($template->id)->get() as $device) {
             $device->services()->updateOrCreate(
@@ -285,9 +284,6 @@ class ServiceTemplateController extends Controller
                 ]
             );
         }
-        $msg = __('Services for Template :name have been updated', ['name' => $template->name]);
-
-        return response($msg, 200);
     }
 
     /**
@@ -297,6 +293,8 @@ class ServiceTemplateController extends Controller
      */
     public function applyAll()
     {
+        $this->authorize('service-template.update');
+
         foreach (ServiceTemplate::all() as $template) {
             $this->apply($template);
         }
@@ -307,17 +305,15 @@ class ServiceTemplateController extends Controller
 
     /**
      * Apply all Service Templates for a device
-     *
-     * @return \Illuminate\Http\RedirectResponse|\Illuminate\Http\Response|\Illuminate\View\View
      */
-    public function applyDeviceAll(int $device_id)
+    public function applyDeviceAll(int $device_id): void
     {
         foreach (ServiceTemplate::all() as $template) {
+            if ($template->type == 'dynamic') {
+                $template->updateDevices();
+            }
             $this->applyDevice($template, $device_id);
         }
-        $msg = __('All Service Templates have been applied to device ' . $device_id);
-
-        return response($msg, 200);
     }
 
     /**
@@ -328,15 +324,17 @@ class ServiceTemplateController extends Controller
      */
     public function apply(ServiceTemplate $template)
     {
+        $this->authorize('update', $template);
+
         if ($template->type == 'dynamic') {
             $template->updateDevices();
         }
-        ServiceTemplateController::applyDevices($template);
-        ServiceTemplateController::applyDeviceGroups($template);
+        $this->applyDevices($template);
+        $this->applyDeviceGroups($template);
 
         // remove any remaining services no longer in the correct device group
         foreach (Device::notInServiceTemplate($template->id)->notInDeviceGroup($template->groups->pluck('id'))->pluck('device_id') as $device_id) {
-            Service::where('device_id', $device_id)->where('service_template_id', $template->id)->delete();
+            $template->services()->where('device_id', $device_id)->delete();
         }
         $msg = __('All Service Templates have been applied');
 
@@ -348,9 +346,8 @@ class ServiceTemplateController extends Controller
      *
      * @param  ServiceTemplate  $template
      * @param  int  $device_id
-     * @return \Illuminate\Http\RedirectResponse|\Illuminate\Http\Response|\Illuminate\View\View
      */
-    public function applyDevice(ServiceTemplate $template, int $device_id)
+    private function applyDevice(ServiceTemplate $template, int $device_id): void
     {
         // Check if the device needs to be added
         foreach (Device::inServiceTemplate($template->id)->where('device_id', $device_id)->get() as $device) {
@@ -370,7 +367,7 @@ class ServiceTemplateController extends Controller
                 ]
             );
 
-            return response('Service template ' . $template->id . ' applied to device ID ' . $device_id, 200);
+            return; // found
         }
 
         foreach (DeviceGroup::inServiceTemplate($template->id)->get() as $device_group) {
@@ -391,16 +388,14 @@ class ServiceTemplateController extends Controller
                     ]
                 );
 
-                return response('Service template ' . $template->id . ' applied to device ID ' . $device_id, 200);
+                return; // found
             }
         }
 
         // remove if this template no longer applies
         foreach (Device::notInServiceTemplate($template->id)->notInDeviceGroup($template->groups->pluck('id'))->where('device_id', $device_id)->pluck('device_id') as $device_id) {
-            Service::where('device_id', $device_id)->where('service_template_id', $template->id)->delete();
+            $template->services()->where('device_id', $device_id)->delete();
         }
-
-        return response('Service template ' . $template->id . ' applied to device ID ' . $device_id, 200);
     }
 
     /**
@@ -411,7 +406,9 @@ class ServiceTemplateController extends Controller
      */
     public function remove(ServiceTemplate $template)
     {
-        Service::where('service_template_id', $template->id)->delete();
+        $this->authorize('update', $template);
+
+        $template->services()->delete();
 
         $msg = __('All Service Templates have been applied');
 
@@ -426,7 +423,9 @@ class ServiceTemplateController extends Controller
      */
     public function destroy(ServiceTemplate $template)
     {
-        Service::where('service_template_id', $template->id)->delete();
+        $this->authorize('delete', $template);
+
+        $template->services()->delete();
         $template->delete();
 
         $msg = __('Service Template :name deleted, Services removed', ['name' => htmlentities($template->name)]);
